+++ Kernel/Modules/AdminNotificationEvent.pm
@@ -99,7 +99,10 @@ sub Run {
         $Self->{LayoutObject}->ChallengeTokenCheck();
 
         my %GetParam;
-        for my $Parameter (qw(ID Name Subject Body Type Charset Comment ValidID Events)) {
+        for my $Parameter (
+            qw(ID Name Subject Body Type Charset Comment ValidID Events ArticleSubjectMatch ArticleBodyMatch ArticleTypeID ArticleSenderTypeID)
+            )
+        {
             $GetParam{$Parameter} = $Self->{ParamObject}->GetParam( Param => $Parameter ) || '';
         }
         for my $Parameter (
@@ -151,12 +154,37 @@ sub Run {
         }
 
         # update
-        my $Ok = $Self->{NotificationEventObject}->NotificationUpdate(
-            %GetParam,
-            Charset => $Self->{LayoutObject}->{UserCharset},
-            Type    => 'text/plain',
-            UserID  => $Self->{UserID},
-        );
+        my $Ok;
+        my $ArticleFilterMissing;
+
+        # checking if article filter exist if necessary
+        if (
+            grep { $_ eq 'ArticleCreate' || $_ eq 'ArticleSend' }
+            @{ $GetParam{Data}->{Events} || [] }
+            )
+        {
+            if (
+                !$GetParam{ArticleTypeID}
+                && !$GetParam{ArticleSenderTypeID}
+                && $GetParam{ArticleSubjectMatch} eq ''
+                && $GetParam{ArticleBodyMatch} eq ''
+                )
+            {
+                $ArticleFilterMissing = 1;
+            }
+        }
+
+        # required Article filter only on ArticleCreate and Article Send event
+        # if isn't selected at least one of the article filter fields, notification isn't updated
+        if ( !$ArticleFilterMissing ) {
+            $Ok = $Self->{NotificationEventObject}->NotificationUpdate(
+                %GetParam,
+                Charset => $Self->{LayoutObject}->{UserCharset},
+                Type    => 'text/plain',
+                UserID  => $Self->{UserID},
+            );
+        }
+
         if ($Ok) {
             $Self->_Overview();
             my $Output = $Self->{LayoutObject}->Header();
@@ -167,7 +195,6 @@ sub Run {
                 Data         => \%Param,
             );
             $Output .= $Self->{LayoutObject}->Footer();
-
             return $Output;
         }
         else {
@@ -177,6 +204,20 @@ sub Run {
                     $GetParam{ $Needed . "ServerError" } = "ServerError";
                 }
             }
+
+            # define ServerError Class atribute if necessary
+            $GetParam{ArticleTypeIDServerError}       = "";
+            $GetParam{ArticleSenderTypeIDServerError} = "";
+            $GetParam{ArticleSubjectMatchServerError} = "";
+            $GetParam{ArticleBodyMatchServerError}    = "";
+
+            if ( $ArticleFilterMissing == 1 ) {
+                $GetParam{ArticleTypeIDServerError}       = "ServerError";
+                $GetParam{ArticleSenderTypeIDServerError} = "ServerError";
+                $GetParam{ArticleSubjectMatchServerError} = "ServerError";
+                $GetParam{ArticleBodyMatchServerError}    = "ServerError";
+            }
+
             my $Output = $Self->{LayoutObject}->Header();
             $Output .= $Self->{LayoutObject}->NavigationBar();
             $Output .= $Self->{LayoutObject}->Notify( Priority => 'Error' );
@@ -223,7 +264,10 @@ sub Run {
         $Self->{LayoutObject}->ChallengeTokenCheck();
 
         my %GetParam;
-        for my $Parameter (qw(Name Subject Body Comment ValidID Events)) {
+        for my $Parameter (
+            qw(Name Subject Body Comment ValidID Events ArticleSubjectMatch ArticleBodyMatch ArticleTypeID ArticleSenderTypeID)
+            )
+        {
             $GetParam{$Parameter} = $Self->{ParamObject}->GetParam( Param => $Parameter ) || '';
         }
         for my $Parameter (
@@ -274,12 +318,36 @@ sub Run {
         }
 
         # add
-        my $ID = $Self->{NotificationEventObject}->NotificationAdd(
-            %GetParam,
-            Charset => $Self->{LayoutObject}->{UserCharset},
-            Type    => 'text/plain',
-            UserID  => $Self->{UserID},
-        );
+        my $ID;
+        my $ArticleFilterMissing;
+
+        # define ServerError Message if necessary
+        if (
+            grep { $_ eq 'ArticleCreate' || $_ eq 'ArticleSend' }
+            @{ $GetParam{Data}->{Events} || [] }
+            )
+        {
+            if (
+                !$GetParam{ArticleTypeID}
+                && !$GetParam{ArticleSenderTypeID}
+                && $GetParam{ArticleSubjectMatch} eq ''
+                && $GetParam{ArticleBodyMatch} eq ''
+                )
+            {
+                $ArticleFilterMissing = 1;
+            }
+        }
+
+        # required Article filter only on ArticleCreate and Article Send event
+        # if isn't selected at least one of the article filter fields, notification isn't added
+        if ( !$ArticleFilterMissing ) {
+            $ID = $Self->{NotificationEventObject}->NotificationAdd(
+                %GetParam,
+                Charset => $Self->{LayoutObject}->{UserCharset},
+                Type    => 'text/plain',
+                UserID  => $Self->{UserID},
+            );
+        }
 
         if ($ID) {
             $Self->_Overview();
@@ -300,6 +368,20 @@ sub Run {
                 if ( $GetParam{$Needed} eq '' ) {
                     $GetParam{ $Needed . "ServerError" } = "ServerError";
                 }
+            }
+
+            # checking if article filter exist if necessary
+            $GetParam{ArticleTypeIDServerError}       = "";
+            $GetParam{ArticleSenderTypeIDServerError} = "";
+            $GetParam{ArticleSubjectMatchServerError} = "";
+            $GetParam{ArticleBodyMatchServerError}    = "";
+
+            if ( $ArticleFilterMissing == 1 )
+            {
+                $GetParam{ArticleTypeIDServerError}       = "ServerError";
+                $GetParam{ArticleSenderTypeIDServerError} = "ServerError";
+                $GetParam{ArticleSubjectMatchServerError} = "ServerError";
+                $GetParam{ArticleBodyMatchServerError}    = "ServerError";
             }
             my $Output = $Self->{LayoutObject}->Header();
             $Output .= $Self->{LayoutObject}->NavigationBar();
@@ -421,6 +503,18 @@ sub _Edit {
     my $EventClass = 'Validate_Required';
     if ( $Param{EventsServerError} ) {
         $EventClass .= ' ' . $Param{EventsServerError};
+    }
+
+    # Set class name for article type...
+    my $ArticleTypeIDClass = '';
+    if ( $Param{ArticleTypeIDServerError} ) {
+        $ArticleTypeIDClass .= ' ' . $Param{ArticleTypeIDServerError};
+    }
+
+    # Set class name for article sender type...
+    my $ArticleSenderTypeIDClass = '';
+    if ( $Param{ArticleSenderTypeIDServerError} ) {
+        $ArticleSenderTypeIDClass .= ' ' . $Param{ArticleSenderTypeIDServerError};
     }
 
     my %RegisteredEvents = $Self->{EventObject}->EventList(
@@ -613,6 +707,7 @@ sub _Edit {
         Data => { $Self->{TicketObject}->ArticleTypeList( Result => 'HASH' ), },
         Name => 'ArticleTypeID',
         SelectedID  => $Param{Data}->{ArticleTypeID},
+        Class       => $ArticleTypeIDClass,
         Size        => 5,
         Multiple    => 1,
         Translation => 1,
@@ -623,6 +718,7 @@ sub _Edit {
         Data => { $Self->{TicketObject}->ArticleSenderTypeList( Result => 'HASH' ), },
         Name => 'ArticleSenderTypeID',
         SelectedID  => $Param{Data}->{ArticleSenderTypeID},
+        Class       => $ArticleSenderTypeIDClass,
         Size        => 5,
         Multiple    => 1,
         Translation => 1,